backend/doly/requirements.txt
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
django-graphql-jwt==0.3.1
graphene_django==2.12.1
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
whitenoise==5.2.0